middleware/auth_service/pyproject.toml
[project]
name = "smarthome-auth-service"
version = "0.1.0"
description = "Authentication middleware microservice for SmartHome"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0,<1.0.0",
    "sqlalchemy[asyncio]>=2.0.36,<3.0.0",
    "asyncpg>=0.29.0,<1.0.0",
    "alembic>=1.14.0,<2.0.0",
    "pydantic-settings>=2.6.0,<3.0.0",
    "python-dotenv>=1.0.1,<2.0.0",
    "prometheus-client>=0.21.0,<1.0.0",
    "uvicorn[standard]>=0.30.0,<1.0.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
